Revert "Feature/speedo chart"
