--- PathDataExercises/package.json
@@ -1,5 +1,0 @@
-{
-  "dependencies": {
-    "echarts": "^5.5.1"
-  }
-}
