--- PathDataExercises/svg-path-ex-1/src/App.tsx
@@ -1,12 +1,0 @@
-import React from 'react';
-import EChartPathEditor from './ex1'
-
-const App = () => {
-  return (
-    <div>
-      <EChartPathEditor />
-    </div>
-  );
-};
-
-export default App;
