+++ custom-speedometer/src/App.tsx
@@ -5,7 +5,7 @@ const App = () => {
 
   return (
     <div>
-      <SpeedoChart min={0} max={100} progress={100} />
+      <SpeedoChart min={0} max={100000} progress={522} />
     </div>
   );
 };
