+++ custom-speedometer/src/SpeedoChart.tsx
@@ -1,21 +1,30 @@
 import React, { useEffect, useRef } from 'react';
 import * as echarts from 'echarts';
 
+
 interface SpeedoChartProps {
   min: number;
   max: number;
   progress: number;
 }
 
-const calculatePercentage = (minVal: number, maxVal: number, progressVal: number): number => {
-  let percentage = ((progressVal - minVal) / (maxVal - minVal)) * 100; // Correctly normalize the progress
+const calculatePercentage = (minVal:number, maxVal:number, progressVal:number): number => {
+
+  // Determine min and max for calculation
+  const min = Math.min(minVal, maxVal);
+  const max = Math.max(minVal, maxVal);
+
+  // Calculate the percentage of progress
+  let percentage = ((progressVal) / (max)) * 100;
   percentage = parseFloat(percentage.toFixed(2));
 
-  // Ensure percentage does not exceed 100% or fall below 0%
-  if (percentage > 100) {
+  // Clamp the percentage between 0 and 100
+  console.log("Percentage: "+percentage);
+  console.log("MIN alue: " +min);
+  console.log("MAX Value: "+max);
+
+  if(percentage>100) {
     percentage = 100;
-  } else if (percentage < 0) {
-    percentage = 0;
   }
 
   return percentage;
@@ -25,68 +34,52 @@ const SpeedoChart: React.FC<SpeedoChartProps> = ({ min, max, progress }) => {
   const chartRef = useRef<HTMLDivElement>(null);
 
   const calculatedData = calculatePercentage(min, max, progress);
-
+  
   useEffect(() => {
     const chart = echarts.init(chartRef.current!);
 
     const options = {
       title: {
-        text: `Progress: ${calculatedData}%`,
+        text: `Custom Progress Bar ${calculatedData}%`,
         left: 'center',
-        top: 'center',
-        textStyle: {
-          fontSize: 16,
-          fontWeight: 'bold',
-        },
+        bottom: -5
       },
       xAxis: {
         type: 'value',
         min: 0,
-        max: 100,
-        show: false,
+        max: 100,  // Max value for the progress bar
+        show: true, // Hide the axis lines
       },
       yAxis: {
         type: 'category',
-        data: [''],
-        show: false,
+        data: ['Progress'], // Label for the progress bar
+        show: false, // Hide the axis labels
       },
-      series: [{
-        type: 'custom',
-        renderItem: (params: any, api: any) => {
-          const startAngle = -Math.PI; // Starting angle for the arc (180 degrees)
-          const endAngle = startAngle + (Math.PI * (calculatedData / 100)); // Ending angle based on progress
-      
-          // Create the Rainbow Arch radiuses
-          const outerRadius = 100;
-          const innerRadius = 80;
-      
-          const cx = api.coord([0, 0])[0]; // Center x
-          const cy = api.coord([0, 0])[1]; // Center y
-      
-          const largeArcFlag = calculatedData > 50 ? 0 : 1; // Set large-arc flag if progress > 50%
-      
-          return {
-            type: 'path',
-            shape: {
-              pathData: `
-                M ${cx + innerRadius * Math.cos(startAngle)} ${cy + innerRadius * Math.sin(startAngle)}
-                A ${innerRadius} ${innerRadius} 0 ${largeArcFlag} 1 
-                  ${cx + innerRadius * Math.cos(endAngle)} ${cy + innerRadius * Math.sin(endAngle)}
-                L ${cx + outerRadius * Math.cos(endAngle)} ${cy + outerRadius * Math.sin(endAngle)}
-                A ${outerRadius} ${outerRadius} 0 ${largeArcFlag} 0 
-                  ${cx + outerRadius * Math.cos(startAngle)} ${cy + outerRadius * Math.sin(startAngle)}
-                Z
-              `,
-            },
-            style: {
-              fill: '#4caf50', // Progress color (green)
-              stroke: '#000', // Outline color
-              lineWidth: 2,
-            },
-          };
+      series: [        
+        // Progress indicator
+        {
+          type: 'custom',
+          renderItem: (params: any, api: any) => {
+            const minPosition = api.coord([min, 0])[0];
+            const progressWidth = api.coord([calculatedData, 0])[0] - minPosition;
+            const barHeight = api.size([0, 1])[1] / 2;
+
+            return {
+              type: 'rect',
+              shape: {
+                x: minPosition,
+                y: params.coordSys.height / 2 - barHeight / 2,
+                width: progressWidth, // Dynamic width based on progress
+                height: 50,
+              },
+              style: {
+                fill: '#4caf50', // Progress color (green)
+              },
+            };
+          },
+          data: [[calculatedData]], // Dynamic progress value
         },
-        data: [{}], // Single data item to trigger renderItem
-      }],      
+      ],
     };
 
     chart.setOption(options);
@@ -96,7 +89,7 @@ const SpeedoChart: React.FC<SpeedoChartProps> = ({ min, max, progress }) => {
     };
   }, [progress]);
 
-  return <div ref={chartRef} style={{ width: '100%', height: '520px' }} />;
+  return <div ref={chartRef} style={{ width: '100%', height: '120px' }} />;
 };
 
 export default SpeedoChart;
